Adding a custom version of the PicocliCommands. This won't be needed until picocli 4.3 has been released.
This fix will allow us to use the auto completion of JLine3 when there is a space between the option name and the option value, example: '-l two'. Right now you would have to type '-l=two'. This wouldn't allow any completion because we'd have to use the absolute list and group names.